tests/wpt/tests/html/canvas/element/path-objects/2d.path.roundrect.3.radii.1.dompoint.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.path.roundrect.3.radii.1.dompoint</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.path.roundrect.3.radii.1.dompoint</h1>
<p class="desc">Verify that when three radii are given to roundRect(), the first radius, specified as a DOMPoint, applies to the top-left corner.</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Verify that when three radii are given to roundRect(), the first radius, specified as a DOMPoint, applies to the top-left corner.");
_addTest(function(canvas, ctx) {

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  ctx.roundRect(0, 0, 100, 50, [new DOMPoint(40, 20), 0, 0]);
  ctx.fillStyle = '#0f0';
  ctx.fill();

  // top-left corner
  _assertPixel(canvas, 20,1, 255,0,0,255);
  _assertPixel(canvas, 41,1, 0,255,0,255);
  _assertPixel(canvas, 1,10, 255,0,0,255);
  _assertPixel(canvas, 1,21, 0,255,0,255);

  // other corners
  _assertPixel(canvas, 98,1, 0,255,0,255);
  _assertPixel(canvas, 98,48, 0,255,0,255);
  _assertPixel(canvas, 1,48, 0,255,0,255);

});
</script>
